Track existing devices for changes
When monitoring devices for changes in order to persist the model, we
not only need to monitor newly added devices, we also need to monitor
devices which already exist in the model right from the start.

Otherwise edits to devices won't be persisted unless a device is also
added or deleted at a later point.

Fixes #22