Update hardware support metadata
While the minimal window size is 360px, this size
doesn't really work well; our default is 480, so
let's recommend that in the metadata.

Remover internet requirement; we require network
connectivity to a local network, otherwise this
app doesn't make much sense, but we never even
talk to the internet at all.