Fix memory management for device row monitoring
Do not use glib::clone to keep a weak reference for the long-running
monitor future: the weak reference gets upgraded to a strong ref the
first time the future is polled, and from that point onwards the future
holds a strong reference to the row, making the whole weak clone a moot
thing.

Instead we explicitly downgrade the row to only pass a weak reference
into the future, and then for every item in the stream explicitly
upgrade to a strong ref and exit the monitoring stream if the upgrade
fails.